+++ CHANGELOG.md
@@ -1,14 +1,45 @@
-#  (2020-10-17)
+# [](https://github.com/augejs/koa-modules/compare/v1.0.2-rc.3...v) (2021-06-08)
 
 
 ### Bug Fixes
 
-* refactor hook util ([7242982](https://github.com/augejs/koa-modules/commit/7242982ae2567dc4ec918d891c4772290c6c6980))
+* add eslint ([47c630e](https://github.com/augejs/koa-modules/commit/47c630ea6ecd66f1604744a79231c870e0c4b1d1))
 
 
 ### Features
 
-* koa proxy add middleware export ([1d7a947](https://github.com/augejs/koa-modules/commit/1d7a947f0aac20ff0323beaa15d81324c0b0c58d))
+* add error handle for validator ([b35c7af](https://github.com/augejs/koa-modules/commit/b35c7af1aee29abf1949eef61e4fda629d70dc5a))
+
+
+
+## [1.0.2-rc.3](https://github.com/augejs/koa-modules/compare/v1.0.2-rc.2...v1.0.2-rc.3) (2021-01-09)
+
+
+### Bug Fixes
+
+* koa error handle ([3c63e65](https://github.com/augejs/koa-modules/commit/3c63e65503da8c7724a8765d8b9a56d46d811e5c))
+
+
+
+## [1.0.2-rc.2](https://github.com/augejs/koa-modules/compare/7242982ae2567dc4ec918d891c4772290c6c6980...v1.0.2-rc.2) (2021-01-06)
+
+
+### Bug Fixes
+
+* fix the version error ([93bc36c](https://github.com/augejs/koa-modules/commit/93bc36c760c4473939c2cab6c4d9e1f11f07e613))
+* fix the version error ([ea0fc55](https://github.com/augejs/koa-modules/commit/ea0fc55b03e8b29b71cd22b74d3a7c2ca7f46e3f))
+* fix the version error ([f70f940](https://github.com/augejs/koa-modules/commit/f70f940d18b67ecb8af1b11fae6a7028806e5ec9))
+* refactor hook util ([7242982](https://github.com/augejs/koa-modules/commit/7242982ae2567dc4ec918d891c4772290c6c6980))
+* update app dir ([ec53faa](https://github.com/augejs/koa-modules/commit/ec53faab14248908ce97a00c337275bb7862b1dc))
+* update app dir ([f779d25](https://github.com/augejs/koa-modules/commit/f779d25358dc81a5d94f20bef72de3ed611e799c))
+* update app dir ([9ded39e](https://github.com/augejs/koa-modules/commit/9ded39e4e6eeede21802aac18435d1d276e14b77))
+* update app dir ([ab7137d](https://github.com/augejs/koa-modules/commit/ab7137df2067acb263d36a6e35b63bf9970ba71f))
+* update app dir ([9e140df](https://github.com/augejs/koa-modules/commit/9e140dfb5601463e4fc94685f31dd4d102c3c60b))
+
+
+### Features
+
+* koa proxy add middleware export ([c0c57d1](https://github.com/augejs/koa-modules/commit/c0c57d1ff8ceb3bc6a94d7f2e5e4defb58a13b0d))
 * release for 1.0.0-rc.2 ([71423aa](https://github.com/augejs/koa-modules/commit/71423aaf6cacb137e24ee36954b7328eedc581cf))
 * release for 1.0.0-rc.3 ([10afdf3](https://github.com/augejs/koa-modules/commit/10afdf3db71b52bf41c7d0a5b2474929db4219ee))